+++ core/src/com/eric/mtd/game/model/actor/combat/enemy/EnemyHumvee.java
@@ -1,11 +1,11 @@
-package com.eric.mtd.game.model.actor.enemy;
+package com.eric.mtd.game.model.actor.combat.enemy;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
-import com.eric.mtd.game.model.actor.GameActor;
+import com.eric.mtd.game.model.actor.combat.CombatActor;
 import com.eric.mtd.game.model.actor.interfaces.IPassiveEnemy;
 import com.eric.mtd.game.model.actor.interfaces.IVehicle;
-import com.eric.mtd.game.service.actorfactory.ActorFactory.GameActorPool;
+import com.eric.mtd.game.service.actorfactory.ActorFactory.CombatActorPool;
 
 /**
  * Represents an Enemy Humvee. A passive enemy.
@@ -26,7 +26,7 @@ public class EnemyHumvee extends Enemy implements IVehicle, IPassiveEnemy {
 	public static final Vector2 GUN_POS = new Vector2(4, 26);
 	public static final Vector2 TEXTURE_SIZE = new Vector2(32, 56);
 
-	public EnemyHumvee(TextureRegion actorRegion, GameActorPool<GameActor> pool) {
+	public EnemyHumvee(TextureRegion actorRegion, CombatActorPool<CombatActor> pool) {
 		super(actorRegion, pool, BODY, TEXTURE_SIZE, GUN_POS, SPEED, HEALTH, ARMOR, ATTACK, ATTACK_SPEED, RANGE);
 	}
