+++ core/src/com/eric/mtd/game/model/actor/combat/enemy/EnemyRocketLauncher.java
@@ -1,12 +1,12 @@
-package com.eric.mtd.game.model.actor.enemy;
+package com.eric.mtd.game.model.actor.combat.enemy;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
-import com.eric.mtd.game.model.actor.GameActor;
+import com.eric.mtd.game.model.actor.combat.CombatActor;
 import com.eric.mtd.game.model.actor.projectile.RPG;
 import com.eric.mtd.game.model.actor.projectile.interfaces.IRPG;
 import com.eric.mtd.game.service.actorfactory.ActorFactory;
-import com.eric.mtd.game.service.actorfactory.ActorFactory.GameActorPool;
+import com.eric.mtd.game.service.actorfactory.ActorFactory.CombatActorPool;
 import com.eric.mtd.util.AudioUtil;
 import com.eric.mtd.util.AudioUtil.ProjectileSound;
 import com.eric.mtd.util.Logger;
@@ -31,7 +31,7 @@ public class EnemyRocketLauncher extends Enemy implements IRPG {
 	public static final Vector2 GUN_POS = new Vector2(4, 26);
 	public static final Vector2 TEXTURE_SIZE = new Vector2(32, 56);
 
-	public EnemyRocketLauncher(TextureRegion[] actorRegions, GameActorPool<GameActor> pool) {
+	public EnemyRocketLauncher(TextureRegion[] actorRegions, CombatActorPool<CombatActor> pool) {
 		super(actorRegions, pool, BODY, TEXTURE_SIZE, GUN_POS, SPEED, HEALTH, ARMOR, ATTACK, ATTACK_SPEED, RANGE);
 	}
