+++ core/src/com/eric/mtd/game/model/actor/combat/enemy/EnemySprinter.java
@@ -1,10 +1,10 @@
-package com.eric.mtd.game.model.actor.enemy;
+package com.eric.mtd.game.model.actor.combat.enemy;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
-import com.eric.mtd.game.model.actor.GameActor;
+import com.eric.mtd.game.model.actor.combat.CombatActor;
 import com.eric.mtd.game.model.actor.interfaces.IPassiveEnemy;
-import com.eric.mtd.game.service.actorfactory.ActorFactory.GameActorPool;
+import com.eric.mtd.game.service.actorfactory.ActorFactory.CombatActorPool;
 
 /**
  * Represents an Enemy Sprint A passive enemy.
@@ -25,7 +25,7 @@ public class EnemySprinter extends Enemy implements IPassiveEnemy {
 	public static final Vector2 GUN_POS = new Vector2(4, 26);
 	public static final Vector2 TEXTURE_SIZE = new Vector2(32, 56);
 
-	public EnemySprinter(TextureRegion[] actorRegions, GameActorPool<GameActor> pool) {
+	public EnemySprinter(TextureRegion[] actorRegions, CombatActorPool<CombatActor> pool) {
 		super(actorRegions, pool, BODY, TEXTURE_SIZE, GUN_POS, SPEED, HEALTH, ARMOR, ATTACK, ATTACK_SPEED, RANGE);
 	}
