+++ core/src/com/eric/mtd/game/model/actor/combat/enemy/EnemyTank.java
@@ -1,4 +1,4 @@
-package com.eric.mtd.game.model.actor.enemy;
+package com.eric.mtd.game.model.actor.combat.enemy;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.Batch;
@@ -7,13 +7,13 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer.ShapeType;
 import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Vector2;
-import com.eric.mtd.game.model.actor.GameActor;
+import com.eric.mtd.game.model.actor.combat.CombatActor;
 import com.eric.mtd.game.model.actor.health.interfaces.IPlatedArmor;
 import com.eric.mtd.game.model.actor.interfaces.IVehicle;
 import com.eric.mtd.game.model.actor.projectile.RPG;
 import com.eric.mtd.game.model.actor.projectile.interfaces.IRPG;
 import com.eric.mtd.game.service.actorfactory.ActorFactory;
-import com.eric.mtd.game.service.actorfactory.ActorFactory.GameActorPool;
+import com.eric.mtd.game.service.actorfactory.ActorFactory.CombatActorPool;
 import com.eric.mtd.util.Logger;
 import com.eric.mtd.util.Resources;
 
@@ -42,7 +42,7 @@ public class EnemyTank extends Enemy implements IPlatedArmor, IRPG, IVehicle {
 	private ShapeRenderer body = Resources.getShapeRenderer();
 	private float bodyRotation; //
 
-	public EnemyTank(TextureRegion tankRegion, TextureRegion turretRegion, GameActorPool<GameActor> pool) {
+	public EnemyTank(TextureRegion tankRegion, TextureRegion turretRegion, CombatActorPool<CombatActor> pool) {
 		super(turretRegion, pool, BODY, TEXTURE_TURRET_SIZE, GUN_POS, SPEED, HEALTH, ARMOR, ATTACK, ATTACK_SPEED, RANGE);
 		this.tankBodyRegion = tankRegion;
 	}
@@ -71,7 +71,7 @@ public class EnemyTank extends Enemy implements IPlatedArmor, IRPG, IVehicle {
 	}
 
 	/**
-	 * Body of the Tank. GameActor/Enemy holds the Body of the turret Which we
+	 * Body of the Tank. CombatActor/Enemy holds the Body of the turret Which we
 	 * don't care about for collision detection.
 	 */
 	@Override
