+++ core/src/com/eric/mtd/game/model/actor/health/HealthBar.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer.ShapeType;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.utils.Pool;
-import com.eric.mtd.game.model.actor.GameActor;
+import com.eric.mtd.game.model.actor.combat.CombatActor;
 import com.eric.mtd.game.service.actorfactory.ActorFactory;
 import com.eric.mtd.util.Logger;
 import com.eric.mtd.util.Resources;
@@ -19,7 +19,7 @@ import com.eric.mtd.util.Resources;
  *
  */
 public class HealthBar extends Actor implements Pool.Poolable {
-	private GameActor actor = null;
+	private CombatActor actor = null;
 	private ShapeRenderer backgroundBar = Resources.getShapeRenderer();;
 	private ShapeRenderer healthBar = Resources.getShapeRenderer();;
 	private ShapeRenderer armorBar = Resources.getShapeRenderer();;
@@ -79,7 +79,7 @@ public class HealthBar extends Actor implements Pool.Poolable {
 		}
 	}
 
-	public void setActor(GameActor actor) {
+	public void setActor(CombatActor actor) {
 		if (Logger.DEBUG)
 			System.out.println("Healthbar: Setting actor");
 		this.actor = actor;
