+++ core/src/com/eric/mtd/game/model/actor/projectile/Bullet.java
@@ -13,7 +13,7 @@ import com.badlogic.gdx.scenes.scene2d.actions.MoveToAction;
 import com.badlogic.gdx.utils.Disposable;
 import com.badlogic.gdx.utils.Pool;
 import com.eric.mtd.game.helper.Damage;
-import com.eric.mtd.game.model.actor.GameActor;
+import com.eric.mtd.game.model.actor.combat.CombatActor;
 import com.eric.mtd.game.service.actorfactory.ActorFactory;
 import com.eric.mtd.game.stage.GameStage;
 import com.eric.mtd.util.Logger;
@@ -28,7 +28,7 @@ import com.eric.mtd.util.Resources;
 public class Bullet extends Actor implements Pool.Poolable {
 	private static final float SPEED = 350f;
 	private ShapeRenderer bullet = Resources.getShapeRenderer();
-	private GameActor target, shooter;
+	private CombatActor target, shooter;
 	private Pool<Bullet> pool;
 	
 	public Bullet(Pool<Bullet> pool){
@@ -45,7 +45,7 @@ public class Bullet extends Actor implements Pool.Poolable {
 	 * @param size
 	 *            - Size of the bullet
 	 */
-	public void initialize(GameActor shooter, GameActor target, Vector2 pos, Vector2 size) {
+	public void initialize(CombatActor shooter, CombatActor target, Vector2 pos, Vector2 size) {
 		this.target = target;
 		this.shooter = shooter;
 		this.setPosition(pos.x, pos.y);
