+++ core/src/com/eric/mtd/game/stage/GameStage.java
@@ -11,7 +11,7 @@ import com.badlogic.gdx.utils.viewport.Viewport;
 import com.eric.mtd.MTDGame;
 import com.eric.mtd.game.model.Player;
 import com.eric.mtd.game.model.actor.ActorGroups;
-import com.eric.mtd.game.model.actor.tower.Tower;
+import com.eric.mtd.game.model.actor.combat.tower.Tower;
 import com.eric.mtd.game.model.level.Level;
 import com.eric.mtd.game.model.level.MTDTiledMapRenderer;
 import com.eric.mtd.game.model.level.Map;
